method selection depending upon the run-time type of this only
An expression such as a+b selects the method to use based solely upon
the run-time type of the first argument (a). When a is null, a null
pointer exception is thrown. When b is null, the same method 
(concatenation) will still be called. The method must be able to handle
null arguments, though.